Added 3rd-Person Camera
The game already has a 3rd-person camera. That is, the camera follows Fang around, and is placed just behind Fang. I also made The Imanus to have Kinematic physics, to see if that prevents him from falling over to the side, since, if enough time passes, he still fell to the side.