You Exit Game When Pressing Esc
The player can now exit the game if they press the Esc key. I've created a Build of the game, and, indeed, you can only exit the game while pressing Esc on the Press Start scene, NOT on any other scene. However, the Build was quite buggy: you start on the Gameplay scene, NOT on the title scene; and only the capsule moves, NOT the medieval cartoon character; and you also moved slowly.