BUGGY: Added RigidBody to Imanus' Capsule
I added 2 empty scripts, which will store the Collision Detection for both the imanus and Fang. Also, I added a RigidBody to the Imanus' capsule. He now damages and kills fang as soon as the capsule touches Fang. However, the capsule now has gravity, and now the Imanus' capsule falls over when falling to the ground.